model/pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "dagshub>=0.6.8",
    "mlflow>=3.10.1",
    "onnxruntime>=1.24.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "skl2onnx>=1.20.0",
    "yfinance>=1.2.0",
]
